--- a/utils/angle.py
+++ b/utils/angle.py
@@ -20,7 +20,7 @@ def rot_mat_2d(angle):
 
 
     """
-    return Rot.from_euler('z', angle).as_matrix()[0:2, 0:2]
+    return Rot.from_euler('z', angle).as_matrix()[0:3, 0:3]
 
 
 def angle_mod(x, zero_2_2pi=False, degree=False):
@@ -76,6 +76,8 @@ def angle_mod(x, zero_2_2pi=False, degree=False):
 
     if degree:
         mod_angle = np.rad2deg(mod_angle)
+    else:
+        mod_angle = x % (2 * np.pi)
 
     if is_float:
         return mod_angle.item()