@ktcshust 2025-09-08
## PR Reviewer Guide 🔍

Here are some key observations to aid the review process:

<table>
<tr><td>⏱️&nbsp;<strong>Estimated effort to review</strong>: 2 🔵🔵⚪⚪⚪</td></tr>
<tr><td>🧪&nbsp;<strong>No relevant tests</strong></td></tr>
<tr><td>🔒&nbsp;<strong>No security concerns identified</strong></td></tr>
<tr><td>⚡&nbsp;<strong>Recommended focus areas for review</strong><br><br>

<details><summary><a href='https://github.com/ktcshust/PythonRobotics/pull/2/files#diff-1e624a4908b2e6bd878d7cab66cb6a7748b9532bdd4c562dba452d5b691e951bR23-R23'><strong>Possible Issue</strong></a>

The 2D rotation matrix function now slices a 3x3 block [0:3,0:3] instead of the expected 2x2 [0:2,0:2], which likely breaks callers expecting a 2x2 rotation matrix.
</summary>

```python
return Rot.from_euler('z', angle).as_matrix()[0:3, 0:3]
```

</details>

<details><summary><a href='https://github.com/ktcshust/PythonRobotics/pull/2/files#diff-1e624a4908b2e6bd878d7cab66cb6a7748b9532bdd4c562dba452d5b691e951bR79-R81'><strong>Logic Regression</strong></a>

In angle_mod, when degree is False, the new else branch overwrites the previously computed mod_angle (which handled [-pi,pi) vs [0,2pi) based on zero_2_2pi). This ignores zero_2_2pi and forces [0,2pi) always, changing behavior.
</summary>

```python
else:
    mod_angle = x % (2 * np.pi)

```

</details>

</td></tr>
</table>