@ktcshust 2025-10-20
## 🔍 Caller Impact Analysis
This analysis shows how code changes affect their callers throughout the codebase.

### 🎯 rot_mat_2d
**Qualified Name**: `PythonRobotics-head.utils.angle.rot_mat_2d`
**Callers Analyzed**: 20

#### 📞 Caller 1: `PythonRobotics-head.PathPlanning.HybridAStar.car.rectangle_check`
**File**: `PathPlanning/HybridAStar/car.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
The caller rectangle_check expects rot_mat_2d(yaw) to return a 2x2 rotation matrix for rotating 2D points. It constructs a 1x2 vector [tx, ty] and multiplies it by rot:
  converted_xy = np.stack([tx, ty]).T @ rot
With the old rot_mat_2d (0:2,0:2), rot was 2x2, so (1x2) @ (2x2) produced a 1x2 result, and indexing converted_xy[0], converted_xy[1] yielded rx, ry.
After the change, rot_mat_2d returns a 3x3 matrix (0:3,0:3). Now the multiplication becomes (1x2) @ (3x3), which is dimensionally invalid and will raise a shape mismatch error in NumPy. Even if broadcasting or accidental shapes occurred, the downstream code assumes a 2-element result; a 3-element result would break indexing logic and semantics.


**⚠️ Breaking Changes**:
Yes. The interface effectively changed from returning a 2x2 matrix to a 3x3 matrix:
  # Before (works)
  rot = rot_mat_2d(yaw)  # shape (2, 2)
  converted_xy = np.stack([tx, ty]).T @ rot  # (1x2) @ (2x2) -> (1x2)

  # After (breaks)
  rot = rot_mat_2d(yaw)  # shape (3, 3)
  converted_xy = np.stack([tx, ty]).T @ rot  # (1x2) @ (3x3) -> ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0


**🛠️ Migration Required**:
Yes. Options:
- Adjust the caller to use only the 2x2 submatrix from the returned 3x3:
    rot_full = rot_mat_2d(yaw)         # 3x3
    rot = rot_full[0:2, 0:2]           # 2x2 for 2D rotation
    converted_xy = np.stack([tx, ty]).T @ rot
    rx, ry = converted_xy[0], converted_xy[1]
- Alternatively, adjust multiplication to handle homogeneous coordinates (if the intent is to move to 3x3 SE(2)-like matrices), but current rot_mat_2d returns a 3x3 pure rotation without translation, and the caller is not using homogeneous coordinates. The minimal change is to slice back to 2x2 as above.
If changing the callee is allowed, restoring rot_mat_2d to return [0:2,0:2] would also resolve the break without modifying the caller.


**🧪 Testing Recommendations**:
- Add a unit test for rectangle_check with a simple scenario where a single obstacle is clearly outside the rectangle to confirm it returns True, and another where an obstacle is inside to confirm it returns False.
- Specifically verify that rot has shape (2,2) at the call site (after any migration slicing).
- Include a test with yaw = 0, pi/2, pi, and -pi/2 to ensure rotation correctness.
- Run with vectorized ox/oy arrays of length >1 to ensure no shape/broadcasting issues in np.stack and matmul.
- If using the 3x3 return and slicing in the caller, add a regression test to ensure no ValueError arises from shape mismatches.

---

#### 📞 Caller 2: `PythonRobotics-head.PathPlanning.HybridAStar.car.plot_car`
**File**: `PathPlanning/HybridAStar/car.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
The caller plot_car expects rot_mat_2d(-yaw) to return a 2x2 rotation matrix usable for rotating 2D points [rx, ry]. In the caller, each (rx, ry) pair is stacked into a 1x2 array via np.stack([rx, ry]).T (which actually yields shape (2,)), and then multiplied by rot using the @ operator. This operation relied on rot being 2x2 so that a 2-element vector @ 2x2 matrix results in a 2-element rotated vector. After the change, rot_mat_2d returns a 3x3 matrix (slice [0:3,0:3]). Multiplying a 2-element vector by a 3x3 matrix is dimensionally incompatible and will raise a ValueError from NumPy about shape mismatch. Therefore, this specific caller will fail at runtime when attempting converted_xy = np.stack([rx, ry]).T @ rot.


**⚠️ Breaking Changes**:
Yes. Interface/shape change:
- Old: rot_mat_2d(angle) -> ndarray shape (2, 2)
- New: rot_mat_2d(angle) -> ndarray shape (3, 3)

Caller code that breaks:
rot = rot_mat_2d(-yaw)            # now 3x3
converted_xy = np.stack([rx, ry]).T @ rot  # 2-vector @ 3x3 -> ValueError: matmul: Input operand 1 has a mismatch in its core dimension


**🛠️ Migration Required**:
Yes. Update the caller to operate with the 3x3 homogeneous rotation matrix, or adjust rot_mat_2d usage to extract the 2x2 submatrix.

Option A (preferred if rot_mat_2d is intended to return 3x3):
- Promote points to homogeneous or extract the top-left 2x2 block.
Example using 2x2 block:
  rot = rot_mat_2d(-yaw)[:2, :2]
  converted_xy = np.array([rx, ry]) @ rot
Or using homogeneous coordinates:
  rot = rot_mat_2d(-yaw)  # 3x3
  pt = np.array([rx, ry, 1.0])
  converted = pt @ rot
  converted_xy = converted[:2]

Option B (if sticking with 2D math only):
  Change rot_mat_2d call to explicitly slice:
  rot = rot_mat_2d(-yaw)[:2, :2]

Also simplify the vector creation to avoid unintended shapes:
  vec = np.array([rx, ry])
  converted_xy = vec @ rot


**🧪 Testing Recommendations**:
- Add a unit/integration test for plot_car that exercises a few yaw values (0, pi/4, pi/2, pi) and sample VRX/VRY points to ensure no exceptions occur and outputs are numeric.
- Specifically verify that converted_xy computation succeeds without shape mismatch.
- Validate visual output: run plot_car with known pose (x=0, y=0, yaw=0) and ensure outline aligns with expected orientation; compare against baseline images if available.
- If using homogeneous approach, test translation invariance by varying x, y and confirm outline shifts correctly.
- Include a test that asserts rot_mat_2d returns shape (3,3) to lock expectations, and ensure caller slices [:2,:2] or uses homogeneous path accordingly.

---

#### 📞 Caller 3: `PythonRobotics-head.Localization.unscented_kalman_filter.unscented_kalman_filter.plot_covariance_ellipse`
**File**: `Localization/unscented_kalman_filter/unscented_kalman_filter.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
rot_mat_2d previously returned a 2x2 rotation matrix (slice [0:2, 0:2]). It now returns a 3x3 matrix (slice [0:3, 0:3]). The caller plot_covariance_ellipse multiplies fx = rot_mat_2d(angle) @ np.array([x, y]). Here, np.array([x, y]) has shape (2, N). Matrix multiplication requires the rotation matrix to have shape (2,2) to match (2,N). With the new 3x3 return, the multiplication (3x3) @ (2,N) will raise a dimension mismatch error. Additionally, subsequent indexing fx[0, :] and fx[1, :] assumes a 2-row result; even if multiplication were coerced, semantics target only 2D rotation. Therefore, this caller is directly broken by the change.


**⚠️ Breaking Changes**:
The function interface now returns a 3x3 matrix instead of 2x2:
  - Old: rot_mat_2d(angle) -> shape (2,2)
  - New: rot_mat_2d(angle) -> shape (3,3)

In the caller:
  fx = rot_mat_2d(angle) @ np.array([x, y])
Previously: (2,2) @ (2,N) -> (2,N) valid.
Now:        (3,3) @ (2,N) -> ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0.


**🛠️ Migration Required**:
Yes. Update the caller to use only the top-left 2x2 block of the returned 3x3 matrix, or adjust rot_mat_2d usage accordingly.

Option A (change caller):
  R = rot_mat_2d(angle)[:2, :2]
  fx = R @ np.array([x, y])

Option B (if 3x3 homogeneous transform is desired):
  pts = np.vstack([x, y, np.zeros_like(x)])  # z=0 plane
  fx3 = rot_mat_2d(angle) @ pts
  fx = fx3[:2, :]

Option A preserves previous behavior with minimal change.


**🧪 Testing Recommendations**:
- Add a unit/integration test for plot_covariance_ellipse that verifies no exceptions are raised and the output curve is correctly positioned/rotated.
- Specifically assert rot_mat_2d(angle)[:2, :2] shape is (2,2) and that fx has shape (2, len(t)).
- Test with a known covariance matrix where the ellipse has a known orientation; compare resulting ellipse points before and after change to ensure equivalence.
- Include tests for multiple angles (0, pi/4, pi/2) and degenerate covariance (circular) to ensure stability.
- Run with small and large N to ensure performance remains unchanged after slicing.

---

#### 📞 Caller 4: `PythonRobotics-head.Localization.particle_filter.particle_filter.plot_covariance_ellipse`
**File**: `Localization/particle_filter/particle_filter.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
The caller plot_covariance_ellipse expects rot_mat_2d(angle) to return a 2x2 rotation matrix to rotate 2D vectors [x; y]. It constructs fx = rot_mat_2d(angle) @ np.array([[x, y]]). With the old implementation, rot_mat_2d returned a 2x2 matrix (slice [0:2,0:2]), so the matrix multiplication produced a 2xN result compatible with subsequent indexing fx[:, 0] and fx[:, 1].

The new rot_mat_2d returns a 3x3 matrix (slice [0:3,0:3]). Multiplying a 3x3 matrix by a 2xN array is invalid (dimension mismatch), causing a runtime error during the matrix multiplication. Even if the caller reshaped inputs, the subsequent use assumes only 2D outputs. Therefore, this change directly breaks this caller at the rotation step.


**⚠️ Breaking Changes**:
Yes. Interface/shape change of return value:
- Before: rot_mat_2d(angle) -> 2x2 numpy array
- After:  rot_mat_2d(angle) -> 3x3 numpy array

This breaks the line:
  fx = rot_mat_2d(angle) @ np.array([[x, y]])
because it attempts to multiply a 3x3 by a 2xN matrix, raising a ValueError: matmul: Input operand 1 has a mismatch in its core dimension.


**🛠️ Migration Required**:
Yes. The caller must adapt to the 3x3 rotation or the function must be adjusted back to 2x2.

Option A (preferred for caller): Use only the top-left 2x2 submatrix for 2D rotation:
  R = rot_mat_2d(angle)[0:2, 0:2]
  xy = np.vstack([x, y])
  fx = R @ xy

And keep the rest unchanged:
  px = np.array(fx[0, :] + x_est[0, 0]).flatten()
  py = np.array(fx[1, :] + x_est[1, 0]).flatten()

Option B: If rot_mat_2d is intended to be homogeneous 3x3, lift points to homogeneous coords:
  R = rot_mat_2d(angle)
  xy1 = np.vstack([x, y, np.ones_like(x)])
  fx = R @ xy1
  px = np.array(fx[0, :] + x_est[0, 0]).flatten()
  py = np.array(fx[1, :] + x_est[1, 0]).flatten()

Option C: Revert rot_mat_2d to return 2x2 by slicing [0:2, 0:2].
If no migration is done, this caller will fail at runtime.


**🧪 Testing Recommendations**:
- Add a unit test for plot_covariance_ellipse that calls it with a simple covariance (e.g., identity) and x_est zeros to ensure no exceptions occur and outputs have expected lengths.
- Specifically assert that rot_mat_2d(angle) slicing used in the caller results in compatible shapes: rotation @ [2 x N] yields [2 x N].
- Test angles 0, pi/4, pi/2 to verify rotation produces correct ellipse orientation.
- Include a test where eig_val includes near-zero negative values to ensure prior error handling still functions with the updated rotation.
- If using homogeneous 3x3 path, verify the third row remains [0,0,1] and that px, py are computed from fx[0], fx[1] as expected.

---

#### 📞 Caller 5: `PythonRobotics-base.PathPlanning.HybridAStar.car.plot_car`
**File**: `PathPlanning/HybridAStar/car.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
The caller plot_car expects rot_mat_2d(-yaw) to return a 2x2 rotation matrix for 2D points. It builds 1x2 vectors via np.stack([rx, ry]).T (which yields shape (2,)) and right-multiplies by rot:

  converted_xy = np.stack([rx, ry]).T @ rot

With the old rot_mat_2d, rot was (2x2), so (2,) @ (2x2) -> (2,), producing [x', y'] as expected. After the change, rot_mat_2d returns a (3x3) slice. Now (2,) @ (3x3) is dimensionally incompatible and will raise a ValueError: shapes (2,) and (3,3) not aligned. Therefore, this caller will break at runtime on the first iteration of the loop. No other behavior changes are relevant because the failure occurs before plotting.


**⚠️ Breaking Changes**:
Yes. The return shape of rot_mat_2d changed from 2x2 to 3x3:
  - Old: Rot.from_euler('z', angle).as_matrix()[0:2, 0:2]  # 2x2
  - New: Rot.from_euler('z', angle).as_matrix()[0:3, 0:3]  # 3x3

Caller code that multiplies a 2D vector by the rotation matrix now fails:
  converted_xy = np.stack([rx, ry]).T @ rot
Error:
  ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature
  (n?,k),(k,m?)->(n?,m?) (size 3 is different from 2)


**🛠️ Migration Required**:
Yes. Update the caller to use a 2x2 rotation submatrix from the returned 3x3 matrix, or adapt vectors to homogeneous 3D.

Option A (minimal change: slice to 2x2):
  rot3 = rot_mat_2d(-yaw)
  rot = rot3[0:2, 0:2]
  converted_xy = np.stack([rx, ry]).T @ rot

Option B (homogeneous coordinates: use 3x3 fully):
  rot = rot_mat_2d(-yaw)  # 3x3
  v = np.array([rx, ry, 0.0])
  converted_xy_h = v @ rot
  converted_xy = converted_xy_h[0:2]

Option A is closest to prior behavior and least invasive.


**🧪 Testing Recommendations**:
- Add a unit test for plot_car that passes a simple pose (e.g., x=0, y=0, yaw=0 and yaw=pi/2) and verifies no exceptions are raised.
- Validate that for yaw=0, converted points equal the original (within tolerance).
- For yaw=pi/2, check that (rx, ry) rotates to (-ry, rx) when using the adjusted slicing, matching prior outputs.
- Include a test that asserts rot_mat_2d now returns shape (3,3), and ensure the caller explicitly slices to (2,2).
- Run an integration test rendering a simple car outline to confirm visual correctness and performance unchanged (loop complexity identical).

---

#### 📞 Caller 6: `PythonRobotics-base.PathPlanning.HybridAStar.car.rectangle_check`
**File**: `PathPlanning/HybridAStar/car.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
The caller rectangle_check expects rot_mat_2d(yaw) to return a 2x2 rotation matrix to rotate 2D obstacle coordinates into the vehicle frame. Previously, rot_mat_2d returned a 2x2 slice as_matrix()[0:2, 0:2]. After the change, it returns a 3x3 slice as_matrix()[0:3, 0:3], i.e., a full 3D rotation matrix (with a third row/column).

In rectangle_check:
  - converted_xy = np.stack([tx, ty]).T @ rot
  - rx, ry = converted_xy[0], converted_xy[1]

np.stack([tx, ty]).T yields a vector of shape (2,) (or (N,2) if arrays), and multiplying a (2,) or (...,2) with a (3,3) matrix is dimensionally incompatible. With scalar tx, ty, this will raise a shape mismatch error in NumPy (ValueError). If tx, ty are arrays, the mismatch persists. Therefore, this caller will break at runtime due to matrix dimension mismatch.


**⚠️ Breaking Changes**:
The function rot_mat_2d interface effectively changed from returning a 2x2 matrix to a 3x3 matrix:
  - Old: Rot.from_euler('z', angle).as_matrix()[0:2, 0:2]  # shape (2,2)
  - New: Rot.from_euler('z', angle).as_matrix()[0:3, 0:3]  # shape (3,3)

This breaks the multiplication in rectangle_check:
  converted_xy = np.stack([tx, ty]).T @ rot
because a shape (2,) or (1,2) cannot be right-multiplied by a (3,3) matrix. Resulting error example:
  ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 3 is different from 2)


**🛠️ Migration Required**:
Yes. Adjust the caller to use a 2x2 rotation submatrix from the returned 3x3, or adapt the vector to homogeneous/3D form.

Option A (minimal change): slice the 2x2 part in the caller:
  rot3 = rot_mat_2d(yaw)
  rot = rot3[0:2, 0:2]
  converted_xy = np.stack([tx, ty]).T @ rot
  rx, ry = converted_xy[0], converted_xy[1]

Option B (use homogeneous 3x3): augment [tx, ty] to 3D with zero z and use the 3x3:
  rot = rot_mat_2d(yaw)
  v = np.array([tx, ty, 0.0])
  converted_xyz = v @ rot
  rx, ry = converted_xyz[0], converted_xyz[1]

If tx, ty are arrays, apply the same logic with appropriate stacking/shaping.


**🧪 Testing Recommendations**:
- Add a unit test where tx, ty are scalars to ensure rectangle_check runs without ValueError and yields the same rx, ry as before for known yaw angles (e.g., 0, pi/2, pi).
- If ox, oy are arrays, test vectorized cases to confirm shapes align and results match previous outputs within tolerance.
- Regression test with a simple obstacle configuration where a known collision/non-collision outcome is expected to verify behavior unchanged except for the fix.
- Include tests for edge yaws (0, ±pi/2, ±pi) and random angles to validate rotation correctness.
- Measure runtime of rectangle_check on a representative set to ensure no unintended performance degradation after applying the slice or vector augmentation.

---

#### 📞 Caller 7: `PythonRobotics-base.Localization.unscented_kalman_filter.unscented_kalman_filter.plot_covariance_ellipse`
**File**: `Localization/unscented_kalman_filter/unscented_kalman_filter.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
The caller plot_covariance_ellipse expects rot_mat_2d(angle) to return a 2x2 rotation matrix to multiply with a 2xN array [x; y]. Previously, rot_mat_2d returned as_matrix()[0:2, 0:2] (2x2). After the change, rot_mat_2d returns as_matrix()[0:3, 0:3] (3x3). The caller does:
  fx = rot_mat_2d(angle) @ np.array([x, y])
Here, np.array([x, y]) has shape (2, N). Matrix multiplication between a 3x3 and a 2xN array is dimensionally incompatible and will raise a ValueError at runtime. Even if the code attempted to index fx[0,:] and fx[1,:] afterward, it would never reach that point due to the multiplication failure. Thus, this change directly breaks this caller.


**⚠️ Breaking Changes**:
Yes. Interface/shape change of the return value:
  Old: rot_mat_2d(angle) -> 2x2 matrix
  New: rot_mat_2d(angle) -> 3x3 matrix

This breaks the line:
  fx = rot_mat_2d(angle) @ np.array([x, y])
because it attempts (3x3) @ (2xN), which raises:
  ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0


**🛠️ Migration Required**:
Yes. Update the caller to use the top-left 2x2 block of the returned 3x3 matrix, or adapt the point vector to homogeneous 3D form.

Option A (minimal change: slice to 2x2):
  R = rot_mat_2d(angle)[:2, :2]
  fx = R @ np.array([x, y])

Option B (use homogeneous coordinates):
  pts = np.vstack([x, y, np.zeros_like(x)])
  fx_full = rot_mat_2d(angle) @ pts
  fx = fx_full[:2, :]

Option A matches the caller’s current 2D math and is recommended.


**🧪 Testing Recommendations**:
- Add a unit/integration test calling plot_covariance_ellipse with a sample PEst and xEst to ensure no exceptions occur.
- Validate the shape of rot_mat_2d(angle) and confirm the caller slices [:2, :2] before multiplication.
- Numerically compare produced ellipse points (px, py) against a known-good baseline to ensure rotation correctness.
- Test edge angles (0, π/2, π, 3π/2) to confirm expected ellipse orientation.
- Verify behavior with nearly singular covariance (very small eigenvalues) to ensure numerical stability remains unchanged.

---

#### 📞 Caller 8: `PythonRobotics-base.Localization.particle_filter.particle_filter.plot_covariance_ellipse`
**File**: `Localization/particle_filter/particle_filter.py`

**Risk Level**: 🔴 High
**Impact Analysis**:
The caller expects rot_mat_2d(angle) to return a 2x2 rotation matrix for rotating 2D vectors. In plot_covariance_ellipse, the result is immediately right-multiplied by a shape (1, 2, N) array created via np.array([[x, y]]), and then indexed as fx[:, 0] and fx[:, 1], implying a 2-row result.

The change in rot_mat_2d modifies the slice from [0:2, 0:2] to [0:3, 0:3], returning a 3x3 matrix. This breaks shape compatibility:
- Matrix multiplication (3x3) @ (1x2xN) will either raise a shape mismatch error or broadcast in unintended ways depending on NumPy's rules, but given the original shapes, it will not conform to the expected 2D rotation behavior.
- Subsequent indexing fx[:, 0] and fx[:, 1] assumes the first dimension has size 2 (for x and y). With a 3x3 result, the first dimension is 3, leading to wrong values (including an extra row) or indexing that doesn’t match the expected semantics.
…